requirements.txt
sacrebleu>=2.0.0
openai
requests
sacrebleu
google-genai
dotenv